Implement CreateInvoice app method
